requirements.txt
ansi2html==1.8.0
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter-dash==0.4.2
jupyter_client==7.4.8
jupyter_core==5.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
plotly==5.11.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
retrying==1.3.4
six==1.16.0
stack-data==0.6.2
tenacity==8.1.0
tornado==6.2
traitlets==5.8.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
